+++ FirebaseCore/Sources/Public/FirebaseCore/FIROptions.h
@@ -81,11 +81,6 @@ NS_SWIFT_NAME(FirebaseOptions)
 @property(nonatomic, copy, nullable) NSString *databaseURL;
 
 /**
- * The URL scheme used to set up Durable Deep Link service.
- */
-@property(nonatomic, copy, nullable) NSString *deepLinkURLScheme DEPRECATED_ATTRIBUTE;
-
-/**
  * The Google Cloud Storage bucket name, e.g. @"abc-xyz-123.storage.firebase.com".
  */
 @property(nonatomic, copy, nullable) NSString *storageBucket;