+++ FirebaseCore/CHANGELOG.md
@@ -1,3 +1,10 @@
+# Unreleased
+- [removed] **Breaking change**: FirebaseDynamicLinks has been removed. See
+  https://firebase.google.com/support/dynamic-links-faq for more info.
+- [removed] **Breaking change**: Removed the `Options.deepLinkURLScheme`
+  property. This API was exclusively used by the Dynamic Links SDK, which
+  has been removed.
+
 # Firebase 11.15.0
 - [fixed] Remove c99 as the required C language standard. (#14950)
 - [deprecated] Deprecated the `Options.deepLinkURLScheme` property. This